Refactor: 모임 패키지 리팩토링
