+++ src/main/java/com/comeon/backend/meeting/presentation/response/MeetingListResponse.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 @Getter
 @AllArgsConstructor
-public class MeetingSummaryResponse {
+public class MeetingListResponse {
 
     private Long meetingId;
     private UserSummaryResponse hostUser;
